RavenDB-21946 - fixing PR comments#1
