+++ test/SlowTests/Issues/RavenDB_21946.cs
@@ -1,9 +1,14 @@
 ﻿using System.Collections.Generic;
+using System.Net.Http;
 using System.Threading.Tasks;
 using FastTests;
-using Raven.Client.Documents.Operations.Configuration;
+using Raven.Client.Documents.Conventions;
+using Raven.Client.Documents.Operations;
+using Raven.Client.Http;
+using Raven.Client.Json.Serialization;
 using Raven.Client.ServerWide;
 using Raven.Client.ServerWide.Operations;
+using Sparrow.Json;
 using Tests.Infrastructure;
 using Xunit;
 using Xunit.Abstractions;
@@ -58,6 +63,39 @@ namespace SlowTests.Issues
 
                 Assert.True(conflictSolver3.ResolveByCollection.TryGetValue("Users", out var scriptResolver3));
                 Assert.NotEqual(scriptResolver2.LastModifiedTime, scriptResolver3.LastModifiedTime);
+            }
+        }
+
+        private sealed class GetConflictSolverConfigurationOperation : IMaintenanceOperation<ConflictSolver>
+        {
+            public RavenCommand<ConflictSolver> GetCommand(DocumentConventions conventions, JsonOperationContext context)
+            {
+                return new GetConflictSolverConfigurationCommand();
+            }
+
+            private sealed class GetConflictSolverConfigurationCommand : RavenCommand<ConflictSolver>
+            {
+                public override bool IsReadRequest => true;
+
+                public override HttpRequestMessage CreateRequest(JsonOperationContext ctx, ServerNode node, out string url)
+                {
+                    url = $"{node.Url}/databases/{node.Database}/replication/conflicts/solver";
+
+                    var request = new HttpRequestMessage
+                    {
+                        Method = HttpMethod.Get
+                    };
+
+                    return request;
+                }
+
+                public override void SetResponse(JsonOperationContext context, BlittableJsonReaderObject response, bool fromCache)
+                {
+                    if (response == null)
+                        return;
+
+                    Result = JsonDeserializationClient.ConflictSolverConfiguration(response);
+                }
             }
         }
     }
